refactor(admin): share the optimistic day-block state between schedule views
WeekView and DayAgendaView each carried the same ~55 lines - the pending set,
the override map, the 1200ms refresh debounce and the reconcile-on-new-events
effect - identical down to the comments. Extract useOptimisticDayBlocks.

The hook also clears the debounce timer on unmount, which neither copy did:
navigating away mid-debounce fired router.refresh() against a view that was
already gone.
