+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tech-support-site",
-  "version": "4.7.16",
+  "version": "4.7.17",
   "private": true,
   "type": "module",
   "scripts": {
